backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
requests>=2.31,<3.0
celery[redis]>=5.3,<6.0
django-celery-beat>=2.5,<3.0
psycopg2-binary>=2.9,<3.0
cryptography>=41.0,<43.0
gunicorn>=21.2,<23.0
python-decouple>=3.8,<4.0
google-generativeai>=0.8.0
google-genai>=0.1.0
groq>=0.4.0
